+++ app/src/main/java/com/oleh/myweatherapp/weather/WeatherServiceImpl.java
@@ -33,7 +33,7 @@ public class WeatherServiceImpl implements WeatherService {
                     rootNode.get("coord").get("lon").asDouble(),
                     rootNode.get("weather").get(0).get("description").asText(),
                     rootNode.get("weather").get(0).get("icon").asText(),
-                    rootNode.get("main").get("temp").asDouble(),
+                    rootNode.get("main").get("temp").asInt(),
                     rootNode.get("main").get("feels_like").asDouble(),
                     rootNode.get("main").get("temp_min").asDouble(),
                     rootNode.get("main").get("temp_max").asDouble(),
